+++ archive_query_log/parsers/url_query.py
@@ -25,7 +25,7 @@ def add_url_query_parser(
         config: Config,
         provider_id: str,
         url_pattern_regex: str | None,
-        priority: int | None,
+        priority: float | None,
         parser_type: UrlQueryParserType,
         parameter: str | None,
         segment: int | None,
